EDGAR -> Cleanup old executor logic
