+++ opendut-edgar/src/service/mod.rs
@@ -4,6 +4,5 @@ pub mod network_interface;
 mod cluster_assignment;
 mod cannelloni_manager;
 mod can_manager;
-mod executor;
 mod vpn;
 mod test_execution;
