+++ opendut-edgar/src/service/test_execution/executor_manager.rs
@@ -1,6 +1,6 @@
 use std::sync::{Arc, Mutex};
 
-use opendut_types::peer::executor::ExecutorDescriptor;
+use opendut_types::peer::{self, executor::ExecutorDescriptor};
 use tokio::sync::watch::{self, Sender};
 use tracing::warn;
 
@@ -19,38 +19,51 @@ impl ExecutorManager {
         }))
     }
 
-    pub fn create_new_executor(&mut self, executor_descriptor: ExecutorDescriptor) {
-        let (tx, rx) = watch::channel(false);
+    pub fn create_new_executors(&mut self, executors: Vec<peer::configuration::Parameter<ExecutorDescriptor>>) {
 
-        match executor_descriptor {
-            ExecutorDescriptor::Executable => warn!("Executing Executable not yet implemented."),
-            ExecutorDescriptor::Container {
-                engine,
-                name,
-                image,
-                volumes: _,
-                devices: _,
-                envs,
-                ports: _,
-                command,
-                args,
-                results_url
-            } => {
-                let container_config = ContainerConfiguration{
-                    name,
+        let executors = executors.into_iter()
+            .filter_map(|executor| { //TODO properly handle Present vs. Absent
+                if matches!(executor.target, peer::configuration::ParameterTarget::Present) {
+                    Some(executor.value)
+                } else {
+                    None
+                }
+            });
+
+        for executor in executors {
+
+            let (tx, rx) = watch::channel(false);
+
+            match executor {
+                ExecutorDescriptor::Executable => warn!("Executing Executable not yet implemented."),
+                ExecutorDescriptor::Container {
                     engine,
+                    name,
                     image,
+                    volumes: _,
+                    devices: _,
+                    envs,
+                    ports: _,
                     command,
                     args,
-                    envs,
-                    results_url,
-                };
-                tokio::spawn(async move {
-                    ContainerManager::new(container_config, rx).start().await;
-                });
+                    results_url
+                } => {
+                    let container_config = ContainerConfiguration{
+                        name,
+                        engine,
+                        image,
+                        command,
+                        args,
+                        envs,
+                        results_url,
+                    };
+                    tokio::spawn(async move {
+                        ContainerManager::new(container_config, rx).start().await;
+                    });
+                }
             }
+            self.tx_termination_channels.push(tx);
         }
-        self.tx_termination_channels.push(tx);
     }
 
     pub fn terminate_executors(&mut self) {
